+++ src/main/java/com/example/sillok_server/domain/post/common/controller/PostController.java
@@ -1,9 +1,11 @@
-package com.example.sillok_server.domain.post.presentation;
+package com.example.sillok_server.domain.post.common.controller;
 
+import com.example.sillok_server.domain.post.common.service.CreatePostService;
+import com.example.sillok_server.domain.post.common.service.QueryPostsByCategoryService;
+import com.example.sillok_server.domain.post.common.service.QueryPostsService;
 import com.example.sillok_server.domain.post.domain.enums.Category;
-import com.example.sillok_server.domain.post.presentation.dto.request.PostRequest;
-import com.example.sillok_server.domain.post.presentation.dto.response.PostResponse;
-import com.example.sillok_server.domain.post.service.*;
+import com.example.sillok_server.domain.post.dto.request.PostRequest;
+import com.example.sillok_server.domain.post.dto.response.PostResponse;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
