+++ src/main/java/com/example/sillok_server/domain/post/common/service/QueryPostsByCategoryService.java
@@ -1,8 +1,8 @@
-package com.example.sillok_server.domain.post.service;
+package com.example.sillok_server.domain.post.common.service;
 
 import com.example.sillok_server.domain.post.domain.enums.Category;
 import com.example.sillok_server.domain.post.domain.repository.PostRepository;
-import com.example.sillok_server.domain.post.presentation.dto.response.PostResponse;
+import com.example.sillok_server.domain.post.dto.response.PostResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
