+++ src/main/java/com/example/sillok_server/domain/post/common/service/QueryPostsService.java
@@ -1,8 +1,7 @@
-package com.example.sillok_server.domain.post.service;
+package com.example.sillok_server.domain.post.common.service;
 
 import com.example.sillok_server.domain.post.domain.repository.PostRepository;
-import com.example.sillok_server.domain.post.presentation.dto.response.PostResponse;
-import com.example.sillok_server.domain.user.facade.UserFacade;
+import com.example.sillok_server.domain.post.dto.response.PostResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -11,15 +10,13 @@ import java.util.List;
 
 @Service
 @RequiredArgsConstructor
-public class AdminQueryPostsService {
+public class QueryPostsService {
 
     private final PostRepository postRepository;
-    private final UserFacade userFacade;
 
     @Transactional(readOnly = true)
     public List<PostResponse> execute() {
-        userFacade.getCurrentUser();
-        return postRepository.findAllByIsApprovedFalse();
+        return postRepository.findAllByIsApprovedTrue();
     }
 
 }
