+++ src/main/java/com/example/sillok_server/domain/post/domain/repository/PostRepositoryCustom.java
@@ -1,7 +1,7 @@
 package com.example.sillok_server.domain.post.domain.repository;
 
 import com.example.sillok_server.domain.post.domain.enums.Category;
-import com.example.sillok_server.domain.post.presentation.dto.response.PostResponse;
+import com.example.sillok_server.domain.post.dto.response.PostResponse;
 
 import java.util.List;
