+++ src/main/java/com/example/sillok_server/domain/post/dto/response/PostResponse.java
@@ -1,4 +1,4 @@
-package com.example.sillok_server.domain.post.presentation.dto.response;
+package com.example.sillok_server.domain.post.dto.response;
 
 import com.example.sillok_server.domain.post.domain.enums.Category;
 import com.querydsl.core.annotations.QueryProjection;
